+++ ChangeLog.md
@@ -1,5 +1,12 @@
 ## Unreleased changes
 
+Major changes:
+
+Other enhancements:
+
+* Added an `allow-newer` config option [#922](https://github.com/commercialhaskell/stack/issues/922) [#770](https://github.com/commercialhaskell/stack/issues/770)
+
+Bug fixes:
 
 ## v0.1.6.0